+++ frontends/rioterm/src/bindings/bindings_winit/mod.rs
@@ -237,7 +237,7 @@ impl From<String> for Action {
             "createwindow" => Some(Action::WindowCreateNew),
             "createtab" => Some(Action::TabCreateNew),
             "closetab" => Some(Action::TabCloseCurrent),
-            "closeothertabs" => Some(Action::TabCloseOthers),
+            "closeunfocusedtabs" => Some(Action::TabCloseUnfocused),
             "openconfigeditor" => Some(Action::ConfigEditor),
             "selectprevtab" => Some(Action::SelectPrevTab),
             "selectnexttab" => Some(Action::SelectNextTab),
@@ -411,7 +411,7 @@ pub enum Action {
     TabCloseCurrent,
 
     /// Close all other tabs (leave only the current tab).
-    TabCloseOthers,
+    TabCloseUnfocused,
 
     /// Toggle fullscreen.
     #[allow(dead_code)]